Add boilerplate script
The templates I've just added give an overview of the action.  The story goes through three states: Growing, FlashUp, and FlashDown.  In each state, entities (a Star and some Motes) react to the passage of time.

To be implemented: What changes the entities undergo, and what graphical objects they will render using that information.  The way that the Graphics object parlays that to our various output devices (projector, LEDs, strobes).